requirements.txt
google-generativeai>=0.8.3
streamlit>=1.39.0
pandas>=2.2.0
matplotlib>=3.9.0
networkx>=3.4.0
jsonschema>=4.23.0
pytest>=8.3.0
python-dotenv>=1.0.0
requests>=2.32.0
plotly>=5.18.0
